requirements.txt
Flask==2.0.2
python-dotenv==0.19.1
slack-sdk==3.10.0
google-api-python-client==2.24.0
pytz==2023.2